Refactor components for better maintainability
Extract reusable components and hooks to reduce duplication:

New components:
- SettingsSection: Wrapper for consistent settings UI
- StateToggleButton: Generic toggle button (used by UnreadToggle, SortToggle)
- ColorPicker/ColorDot: Reusable color selection UI
- SubscriptionItem: Extracted from Sidebar
- EntryContentRenderer: Memoized content rendering
- EntryContentStates: Loading/error states

New hooks and types:
- useFormMessages: Form error/success message handling with auto-clear
- Tag type centralized in @/lib/types
- useNarrationTypes: Shared narration hook types

Updated components:
- TagManagement: Uses SettingsSection, ColorPicker, useFormMessages
- LinkedAccounts: Uses SettingsSection, useFormMessages
- UnreadToggle/SortToggle: Uses StateToggleButton
- EditSubscriptionDialog: Uses centralized Tag type, ColorDot
- Sidebar: Extracted SubscriptionItem, uses ColorDot
- EntryContentBody: Split into focused helper files

Line count reductions:
- TagManagement: 510 → ~300 lines
- Sidebar: 470 → ~380 lines
- EntryContentBody: 759 → ~550 lines
