+++ src/components/ui/index.ts
@@ -27,6 +27,8 @@ export {
   StatusCard,
 } from "./card";
 export { NavLink, NavLinkWithIcon } from "./nav-link";
+export { ColorPicker, ColorDot } from "./color-picker";
+export { StateToggleButton } from "./state-toggle-button";
 export {
   IconButton,
   // Navigation icons
