+++ src/lib/hooks/index.ts
@@ -29,3 +29,5 @@ export {
   type InfiniteScrollConfig,
   type UseInfiniteScrollConfigOptions,
 } from "./useInfiniteScrollConfig";
+
+export { useFormMessages } from "./useFormMessages";
