+++ src/lib/types/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Type Exports
+ *
+ * Centralized exports for shared type definitions.
+ */
+
+export { type Tag, TAG_COLORS, DEFAULT_TAG_COLOR } from "./tag";
